Переделывает форму создания новой точки маршрута
