+++ src/model/destinations-model.js
@@ -3,7 +3,7 @@ import {generateCity} from '../mock/destination-city-mock.js';
 export default class CityModel {
   #cities = Array.from({length: 3}, generateCity);
 
-  get cities() {
+  get tripCities() {
     return this.#cities;
   }
 }
