+++ src/view/wrapper-content-container-view.js
@@ -2,12 +2,11 @@ import AbstractView from '../framework/view/abstract-view.js';
 
 const createWrapperContentContainerTemplate = () => (
   `<li class="trip-events__item">
-
   </li>`
 );
 
 
-export default class WrapperContentContainerView extends AbstractView{
+export default class ContentContainerItemView extends AbstractView{
   get template() {
     return createWrapperContentContainerTemplate();
   }
