--- src/view/wrapper-destination-view.js
@@ -1,12 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-const createDestinationWrapperTemplate = () => (
-  `<section class="event__details">
-  </section>`
-);
-
-export default class TripDestinationWrapperView extends AbstractView{
-  get template() {
-    return createDestinationWrapperTemplate();
-  }
-}
