--- src/view/wrapper-form-content-container-view.js
@@ -1,12 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-const createWrapperFormContentContainerTemplate = () => (
-  `<form class="event event--edit" action="#" method="post">
-  </form>`
-);
-
-export default class WrapperFormContentContainerView extends AbstractView{
-  get template() {
-    return createWrapperFormContentContainerTemplate();
-  }
-}
